platform agnostic compilation

+++ include/core/math/constants.h
@@ -1,4 +1,4 @@
 #pragma once
 
-#define PI 3.14159265359f
-#define PI2 (PI * 2.0f)
+#define PI (float)3.14159265359f
+#define PI2 (float)(PI * 2.0f)
